proper api
